Module G.Ref

val list : t -> (Reference.t * Hash.t) list Lwt.t

list state returns an associated list between reference and its bind hash.

val mem : t -> Reference.t -> bool Lwt.t

mem state ref returns true iff ref exists in state, otherwise returns false.

val read : t -> Reference.t -> (Reference.head_contents, error) Stdlib.result Lwt.t

read state reference returns the value contains in the reference reference (available in the git repository state).

val resolve : t -> Reference.t -> (Hash.t, error) Stdlib.result Lwt.t

resolve state reference returns endpoint of reference (available in the git repository state).

val write : t -> Reference.t -> Reference.head_contents -> (unit, error) Stdlib.result Lwt.t

write state reference value writes the value value in the mutable representation of the reference in the git repository state.

val remove : t -> Reference.t -> (unit, error) Stdlib.result Lwt.t

remove state reference removes the reference reference from the git repository state.